塔防游戏1.0
修改bug
添加射速控制代码
修改子弹飞行速度，和子弹散布

+++ std_tfyx/std_tfyx/chicken.h
@@ -43,7 +43,7 @@ public:
         return position;
     }
 
-    void on_update(int delta)
+    void on_update(float delta)
     {
         if (valid)
         {
